add bench
