+++ rust-enjoyer/Cargo.toml
@@ -3,8 +3,17 @@ name = "rust-enjoyer"
 version = "0.1.0"
 edition = "2021"
 
+
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+criterion = "0.5.1"
 num_cpus = "1.16.0"
 rayon = "1.8.0"
+
+[profile.release]
+opt-level = 3
+
+[[bench]]
+name = "avx_benchmark"
+harness = false
